requirements.txt
Django==3.2.11
gunicorn==20.1.0
opencv-python-headless==4.5.5.64
Pillow==9.1.0
psycopg2==2.9.3
python-decouple==3.5